Add files via upload
Welcome to the  e-Signature App, a user-friendly and versatile solution crafted with HTML, CSS, and JavaScript to streamline the process of capturing  electronic signatures. This project offers a seamless and responsive interface, allowing users to effortlessly draw their signatures on a canvas using mouse or touch input. The elegant design, characterized by a stylish color scheme and the Poppins font from Google Fonts, provides a visually pleasing experience.